Refine sub_pages api and add max_depth option
Instead of passing either a string or a regular expression as the first
argument, require the first argument to be a string of the relative path
whose sub pages we're interested in. A regular expression can now be
passed to the "filter" option instead to further filter the results.

Change the default sort to be first by depth and then alphabetical. Also
got rid of the ability to pass a sort block. The default sorting should
cover > 90% of use cases, and the ability to pass that block just
removes having to type ".sort_by" in front of the block they would be
passing. I think the code is clearer forcing them to type actually use
the "sort_by" method.

Add a "max_depth" option to ignore results that are deeper than the
value passed to max_depth.